@@ -9,7 +9,7 @@ Bundler.require(*Rails.groups)
 module Xronos
   class Application < Rails::Application
     # Initialize configuration defaults for originally generated Rails version.
-    config.load_defaults 5.2
+    config.load_defaults 7.0
 
 		config.assets.precompile += %w(*.png *.jpg *.jpeg *.gif)
 
